Revert "push action to replay buffer after its subtree"
This reverts commit c94e601c6d72ab714150caf020fea9cde9ce4295.

+++ src/engine/DQNReplayBuffer.h
@@ -163,9 +163,9 @@ public:
     torch::Tensor getRewards();
     torch::Tensor getNextStates();
     torch::Tensor getDones();
-    float getDemoFactor() const;
-    float getEpsilonDemo() const;
-    float getEpsilonAgent() const;
+    long getDemoFactor() const;
+    long getEpsilonDemo() const;
+    long getEpsilonAgent() const;
     long getEpsilon() const;
